Saving Data to Local Device File System

+++ XamarinCapstoneProj/XamarinCapstoneProj/Models/Products.cs
@@ -8,7 +8,8 @@ namespace XamarinCapstoneProj.Models
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public float Price { get; set; }
+        public Decimal Price { get; set; }
         public string Description { get; set; }
+
     }
 }
